requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
neo4j==5.27.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
streamlit==1.41.1
requests==2.32.3
